Only send out jobs as previous ones come back
This fixes the bug where new commands will
be run even after an error has been detected

+++ job_linux.go
@@ -14,7 +14,7 @@ func createJob(cmdStr string) job {
 
 // stop a running job - no op if not running yet or already dead
 func (job job) stop() {
-	if job.cmd.Process != nil { // we can only do this if the process exists
+	if job.cmd != nil && job.cmd.Process != nil { // we can only do this if the command and process exists
 		syscall.Kill(-job.cmd.Process.Pid, syscall.SIGTERM) // take down the process group
 	}
 }
